refactor: Controller 통합
