+++ src/main/java/lotto/AppConfig.java
@@ -1,9 +1,8 @@
 package lotto;
 
-import lotto.handler.LottoCheckController;
-import lotto.handler.LottoGenerateController;
-import lotto.handler.LottoResultController;
+import lotto.handler.LottoController;
 import lotto.handler.RandomValueGenerator;
+import lotto.service.LottoService;
 import lotto.view.InputValidator;
 import lotto.view.InputView;
 import lotto.view.OutputView;
@@ -22,19 +21,15 @@ public class AppConfig {
         return new OutputView();
     }
 
-    public LottoGenerateController lottoGenerateController() {
-        return new LottoGenerateController(randomValueGenerator(), inputView(), outputView());
+    public LottoController lottoController() {
+        return new LottoController(inputView(), outputView(), lottoService());
+    }
+
+    public LottoService lottoService() {
+        return new LottoService(randomValueGenerator());
     }
 
     public RandomValueGenerator randomValueGenerator() {
         return new RandomValueGenerator();
-    }
-
-    public LottoCheckController lottoCheckController() {
-        return new LottoCheckController(inputView(), outputView());
-    }
-
-    public LottoResultController lottoResultController() {
-        return new LottoResultController(outputView());
     }
 }
