+++ src/main/java/lotto/service/RandomValueGenerator.java
@@ -1,4 +1,4 @@
-package lotto.handler;
+package lotto.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
