+++ src/test/java/lotto/controller/LottoGenerateControllerTest.java
@@ -1,0 +1,7 @@
+package lotto.controller;
+
+import org.junit.jupiter.api.Test;
+
+class LottoGenerateControllerTest {
+
+}
